For SDT function calls, pick fewer errors if possible
Summary:
When we check calls to SDT functions, we attempt two checks: one using the static parameters, and then again using like- versions of the parameters. If *both* checks fail, we show errors for the "static" check. Unfortunately, if there is a genuine error (e.g. a bool field in a shape instead of a float) then we show this error *as well as* a slew of errors caused by like types not matching non-like types. These latter errors typically don't show up in the "dynamic" check because we are already matching against like types.

Ideally, we should use a heuristic that prefers non-dynamic-related errors. This involves a bit more rejigging of the error API. For now, we simply *count* the errors, and present either the static or dynamic side, whichever has the fewest errors. As you can see from tests, it's usually a good choice.

Reviewed By: madgen

Differential Revision: D46220517

fbshipit-source-id: 04450ee15733f2473d8bf87c462773eef4cca9a4