Initial syntax for typed local variables
Summary:
As an experimental feature, add syntax for local variables that have explicitly declared types.

The syntax is a new statement form: `let` local_var `:` type `=` expression`;`. For example,
```
let $x : int = 1;
let $y : vec<num> = vec[0, 1];
let $z : string = f($x);
```

The experimental feature is named `typed_local_variables`. It emits the same bytecode as a plain assignment statement `$x = 1;`.

The intention is that the type checker will ensure that all assignments are consistent with the given type, but that will be implemented and explained in future commits. It is an open question whether the type, when enforceable, should be enforced at runtime.

At the moment the syntax is not as flexible as it will need to be wrt `await`, `readonly`, etc. We can build that later.

Reviewed By: andrewjkennedy

Differential Revision: D46111299

fbshipit-source-id: c93fe2051480d98471fc44b3dd1ca1f05e33ac70