6/19 - remove test_non_blocking
Summary:
[Diffstack overview: my plan is (1) one by one, port all the tests over to ide_standalone, doing them gradually because different ones require different levels of detail. (2) once that is done, hard-code test_lsp.py to always use ide_standalone, and clean it up. This diff is part of phase 1.]

The "test_non_blocking" test was introduced in D7923548 (May 2018) to prove that hh_server is able to interrupt an ongoing typecheck in order to handle a priority request from the IDE over the persistent connection.

None of this has any meaning any longer, now that we're switch to ide_standalone. I'm therefore deleting the test.

Reviewed By: mheiber

Differential Revision: D46214004

fbshipit-source-id: 211d2c8343d0184afaeaed5d282fcca4bfb7c897