all tests write serverless+standalone to hh.conf
Summary:
[Diffstack overview: my plan is (1) one by one, port all the tests over to ide_standalone, doing them gradually because different ones require different levels of detail. (2) once that is done, hard-code test_lsp.py to always use ide_standalone, and clean it up. This diff is part of phase 2.]

All the tests in test_lsp.py start exactly the same way:
```
        variables = dict(
            self.prepare_serverless_ide_environment(use_standalone_ide=True)
        )
```
It's not clear from the function or parameter name, but the only thing this function is doing is
1. writing hh.conf with flag "ide_serverless=true; ide_standalone={use_standalone_ide}"
2. generating a naming-table and writing it to disk

We don't need this level of customization. I've already changed every single test to use ide_serverless and ide_standalone.

This diff therefore removes the parameters, and hardcodes them always to write hh.conf with "ide_serverless=true; ide_standalone=true"

Reviewed By: mheiber

Differential Revision: D46245730

fbshipit-source-id: d1577394824987b1a6636ff0c84a8a999bb2ce36