12/19 - remove test_hh_server_status_diagnostic
Summary:
[Diffstack overview: my plan is (1) one by one, port all the tests over to ide_standalone, doing them gradually because different ones require different levels of detail. (2) once that is done, hard-code test_lsp.py to always use ide_standalone, and clean it up. This diff is part of phase 1.]

This diff concerns `test_hh_server_status_diagnostic`. Previously, when clientLsp used a persistent connection to hh_server in order to power squiggles, what should it do when that connection was down for whatever reason? Answer, somewhat goofily, it displayed that status as a squiggle. (similar to what arc lint does when hh_server is down).

All of that's gone now: ide_standalone displays hh_server status in the status bar, and has a much richer display. This test no longer serves any purpose, so I'm deleting it.

Reviewed By: mheiber

Differential Revision: D46215786

fbshipit-source-id: 66b0f2f007b8d1e876af696be92c488f2a696204