Refactor HackC changes to ensure only functions in list can be renamed
Summary:
This diff makes it so we don't assume we know anything about what functions we include in the renamable function allowlist in our config. We say its renamable if it appears in the list, regardless if its a builtin.

My initial assumption was that we were going to not include builtins in our renamable function allowlist. I now realize we rename builtins quite a bit, and I should instead start by adding any function we currently rename to the allowlist, and not assume what may be in there.

Reviewed By: aorenste

Differential Revision: D46269227

fbshipit-source-id: a0d7b598bc1d6223a53821883107cfa4a3d8e987