Fixing gcc build
Summary:
X-link: https://github.com/facebook/folly/pull/2009

Apparently this change caused some warning issues on gcc.
There was a shadow warning - that's easy.
The "loosing attributes" warning is a bit trickier, but if you look for it, you will see it disabled everywhere.

It's basically saying that some attributes (like alignas) will be ignored. And __m128 are done with attributes, though these ones are not ignored.

Alternative is to wrap intrinsics but that's too much.

Reviewed By: Orvid

Differential Revision: D46298812

fbshipit-source-id: bf2304704cd9a9b335cf8e7a9f4d3cef6d61ae24