Allow multiple type constraints for type aliases
Summary:
This diff adds ability for type aliases to have multiple type constraints that are meant to represent the union type for type aliases.

This diff threads this all the way through to `PreTypeAlias` and leaves it at converting to the request specific runtime representation of type alias which is `TypeAlias`. Currently we only have one type, so this conversion selects the first type from the list. The following diffs will add this support to `TypeAlias` as well as everywhere we do type enforcement.

Reviewed By: aorenste

Differential Revision: D46200472

fbshipit-source-id: bfd2113894db687c33ac5693603b8cab48fa2cbb