list "static" as reserved name
Summary:
aorenste suggested that "static" be added to the list of reserved names.

I did that with this diff. As proved by the absence of any change in tests, this change has no effect. The reason is that in all type declarations, the name "static" means the line doesn't even get parsed as a declaration (hence we never come to the reserved-name test).
```
// "a name is expected here (Parsing[1002])"
interface Static {}
enum Static : int {}
enum class Static : int {}
case type Static = int;

// "a name is expected here (Parsing[1002])"
// "encountered unexpected text `Static`, was expecting a type hint. (Parsing[1002])"
// "hack does not support top level statements (Parsing[1002])"
// "a semicolon is epxected here (Parsing[1002])"
type Static = int;
newtype Static = int;
newctx Static as [defaults];
module newtype Static as int = int;
```
So, whether or not we include "static" in the list of reserved type names, it has no effect.

Should we add it to the list? Argument in favor is if someone introduces a new type-like construct to the language, but one which does parse it as a name. (If we add it to the list, are the other names we should also add to the list? other modifiers?)

Reviewed By: aorenste

Differential Revision: D45704875

fbshipit-source-id: 3a8e26533d5246f202fc733850446ec7932a7a68