remove a tany
Summary:
Remove the Tany that was generated when looking up an undefined local variable and use `nothing` instead. If we try to look up a missing variable the runtime will halt, so it's ok to use `nothing` in subsequent code.

For array assignments, `$x[e1] = e2`, only set `$x` in the local environment if it wasn't there before to avoid cascading errors (in this case it will be set to `nothing`).

Also refactor typing.ml a bit to change a polymorphic variant into a normal datatype

Reviewed By: madgen

Differential Revision: D46321735

fbshipit-source-id: bbee5631be38578932efd0d784b4b46defebd7b7