Add walkfp command (mini version of walkstk)
Summary:
Adds `walkfp` command. Some fixes to `walkstk` command, and lookup for lowptr mode.

I saw issues with calling `Cast()` on the rip value we create when given an address from the user, so needed to use `unsigned_cast`.

The "<unknown>" frame in the JITed example is a unique stub, which I can add support for displaying later (verified by running LLDB with `env TRACE=ustubs:2` and viewing the corresponding address in `/tmp/hphp.log`)

Reviewed By: oulgen

Differential Revision: D46302289

fbshipit-source-id: 8f8e240179037effd35397a5e40143f2b2c71d4c