portability: handle "" correctly in PosixSpawn
Summary:
Windows handling of argument is weird, very weird in fact. The stackoverflow
article paints a very detailled picture for the interested. One corner case is
for empty strings. It looks like this would get translated to """""" which but
once unexpanded, this would be passed to applications as """. Instead, """"
should be passed, which is correctly unexpanded.

Reviewed By: genevievehelsel

Differential Revision: D46338699

fbshipit-source-id: 3b541354269c1a56737a55cc7598e847d0cc4919