"add null-safe get" code action
Summary:
When `$a->prop` fails due to left-hand-side that might be null, offer the user to replace with `$a?->prop`.

**Choices**

- I used the term "null-safe-get" because [the `?->` operator does not seem to be named in our docs](https://docs.hhvm.com/hack/expressions-and-operators/member-selection). And in our AST it's just `get` with a flag.
- We do not offer the code action when the receiver and the rhs of `$obj->foo` are on different lines. This is because (afaict) we have no way of finding where the `->` is in such cases, due to limitations in the tast representation.

https://pxl.cl/2HGXF

**Prior Art**

Inspiration: Flow. See this screengrab of a similar refactor for Flow: https://pxl.cl/2GnC5

TypeScript doesn't have a quickfix for this error, instead it has really nice autocomplete:

https://pxl.cl/2GnCN

T152879642 is to have similar completions for Hack

Reviewed By: enetsee

Differential Revision: D45698461

fbshipit-source-id: 13774a612b36f8f93d755807a3c2b16ae9916004