remove 'keep_errors'
Summary:
The flag `keep_errors` is only ever set to true. I'm therefore just deleting the flag entirely, to keep things simpler.

I'm particularly happy to delete the flag because of this code:
```
let process_scour_comments (env : env) (sc : Scoured_comments.t) =
  List.iter sc.sc_error_pos ~f:(fun pos ->
      Errors.add_error Parsing_error.(to_user_error @@ Fixme_format pos));
  List.iter sc.sc_bad_ignore_pos ~f:(fun pos ->
      Errors.add_error Parsing_error.(to_user_error @@ Hh_ignore_comment pos));
  if env.keep_errors then (
    Fixme_provider.provide_disallowed_fixmes env.file sc.sc_misuses;
    if env.quick_mode then
      Fixme_provider.provide_decl_hh_fixmes env.file sc.sc_fixmes
    else
      Fixme_provider.provide_hh_fixmes env.file sc.sc_fixmes
  )
```
I don't like that the (global mutable) fixme heap is populated as a side effect of parsing the AST, and I doubly don't like it that this side effect is controlled by a flag. Someone might accidentally set the flag to false, get an AST which they cache, and then subsequent users might consume that cached AST without realizing that they're missing fixmes.

(the ultimate fix is to make fixmes no longer be a global mutable heap. I hope that the simplification in this diff is a small step along the way to that end.)

Reviewed By: edwinsmith

Differential Revision: D46417852

fbshipit-source-id: d9f2fdeaf31407b2a0a0c1b2892821d31f6ff1e7