remove 'disable_global_state_mutation'
Summary:
Thanks to the previous diff, we only ever provide the constant `false` for this flag. I'm therefore removing the flag entirely.

I don't like this flag because of this code:
```
let process_scour_comments (env : env) (sc : Scoured_comments.t) =
  List.iter sc.sc_error_pos ~f:(fun pos ->
      Errors.add_error Parsing_error.(to_user_error @@ Fixme_format pos));
  List.iter sc.sc_bad_ignore_pos ~f:(fun pos ->
      Errors.add_error Parsing_error.(to_user_error @@ Hh_ignore_comment pos));
  if (not env.disable_global_state_mutation) && env.keep_errors then (
    Fixme_provider.provide_disallowed_fixmes env.file sc.sc_misuses;
    if env.quick_mode then
      Fixme_provider.provide_decl_hh_fixmes env.file sc.sc_fixmes
    else
      Fixme_provider.provide_hh_fixmes env.file sc.sc_fixmes
  )
```

If this flag is set to true, then we won't populate fixmes. That's a dangerous thing to do and will break other invariants. I'd like the code to be simple and straightforward and have it always populate fixmes.

Reviewed By: edwinsmith

Differential Revision: D46417851

fbshipit-source-id: 1de80f87f57985209617d4dfa7f0cfca8659e08f