keep_errors should be true
Summary:
This is the only place that sets `keep_errors=false` -- the `keep_errors` flag is used to control whether the act of parsing populates (as a side effect) the global mutable FIXMEs store.

I think this is too dangerous! What if we accidentally parse the file while doing fileOutline, and end up caching the AST without the fixmes, and then we don't have the fixmes in future when we need them?

There's also no benefit I can see to even setting `keep_errors=false`. Why not simply keep errors always?

This diff changes it to `keep_errors=true`.

Reviewed By: edwinsmith

Differential Revision: D46417853

fbshipit-source-id: ceae95225ccaf5214ecb829d2cfeda3dfded6471