Do not consider Enter as an UnconditionalBranch
Summary:
If a call is intercepted, `Enter` does not jump to `targetpc`. CTI needs to use an indirect jump after interpreting `Enter` bytecode.
If the call is *not* intercepted (common case), `Enter` jumps to the `targetpc`, which is always the first instruction of the function, therefore `lookup_cti` does not need to take the slow path.

Reviewed By: ricklavoie

Differential Revision: D46483119

fbshipit-source-id: 4f532d4e8210d3f0d806e39f2f425958e2fbc619