Fix supportdyn wrapping under is/as subexpression, and reset default OCaml elab for hhstc
Summary:
With the OCaml elaborator (default in hh), under implicit pessimisation, `supportdyn` does not get introduced into the subexpression `e` in `e as t` and `e is t`. With the Rust elaborator (default for `hh_single_type_check`), the elaboration is correct.

As we haven't yet rolled out the Rust elaborator, let's revert the default in `hh_single_type_check` to OCaml. (You can use the Rust elaborator with the `--rust-elab` option.)

Thanks to enetsee for the fix to elaboration.

Differential Revision: D46483566

fbshipit-source-id: c4dd96b4835a137dbe831d251dfa7949f110a26f