Set permissions on prefetched db.
Summary:
Sandbox testing revealed some socket/permission bugs.

First, we need to use the configured watchman socket - fixed.

Next, open question: Should we prefetch when Autoload.DB.CanCreate is false?
HHVM currently only tries to fixup the permissions + group when CanCreate=true.

It turns out that the InitializeFacts www script just does chmod(0666) on the file
it downloaded. HHVM will open it, but then not try to set the configured perms or gid,
but it seems to work because 0666 is very permissive.

The purpose of CanCreate was to prevent an expensive full-sync,
which the prefetch avoids.

Reviewed By: aorenste

Differential Revision: D46420503

fbshipit-source-id: 76377fbc4deacc55b56140d1a2b09f2e7e0665f2