Reimplement MapPatch::put
Summary:
Currently `MapPatch::put(...)` is implemented with a new PatchOp: `PatchOp::Put`. However, we can also simplify this and implement it with the existing patch op.

This diff also re-wrote merge unit-test so that it doesn't check the content in `toThrift()` directly. Instead, it checks whether the merged patch has the same behavior as the original patches. This makes the testing more robust.

Reviewed By: iahs

Differential Revision: D46433788

fbshipit-source-id: 72b26a96bef3d88deba126a641d839e7b655be51