Revert D46433788: Multisect successfully blamed D46433788 for test or build failures
Summary:
This diff is reverting D46433788
Depends on D46515787
D46433788: [Thrift][Patch] Reimplement MapPatch::put  by Mizuchi has been identified to be causing the following test or build failures:

Tests affected:
- [icsp/lib/test:codec_tests - ResourceCodecTest.testTenantRoundTrip](https://www.internalfb.com/intern/test/281475051688576/)

Here's the Multisect link:
https://www.internalfb.com/multisect/2211554
Here are the tasks that are relevant to this breakage:

We're generating a revert to back out the changes in this diff, please note the backout may land if someone accepts it.

If you believe this diff has been generated in error you may Commandeer and Abandon it.

Reviewed By: Mizuchi

Differential Revision: D46515798

fbshipit-source-id: cdeadb552287605129b432c2b863b3de051bfdf1