let HazptrObserver accept any hazptr_domain
Summary:
It may occasionally be useful to use a `HazptrObserver` with a domain different from the default domain. Permit this, but default to the default domain and keep the existing interface working.

As an extra, permit any instance of the `hazptr_domain` template. This requires adding the atomic template as a template param.

Reviewed By: Gownta, ot

Differential Revision: D46418921

fbshipit-source-id: 8155974a46faf1d758d54718bddff84971d7394d