Require Case Type Variants to be Recoverable
Summary:
# Context
Case Types up to this point allow any union to be written. We want to limit the unions that can be assigned to a case type based on various properties. The first property is that each variant type of a case type should be recoverable via a type test based on that type.

The notion of being recoverable is tied closely to how refinements are handled. A type `T` is recoverable for a case type `CT` if:
```
(case type CT = V) ∈ Γ
Γ ⊢ T <: V
Γ ⊢ refine(V, T) <: T
-----------------------
Γ ⊢ recoverable(CT, T)
```

We want each variant of a case type to be recoverable. If it is not, we will report an error.

# Implementation
This diff chooses to reuse the existing refinement logic present in the type checker to define the function `refine(V, T)` above. It does so by replacing every type argument during localization with the wildcard `_` and then calling `Typing.refine_and_simplify_intersection`. This is the function that is called when we do refinement based off of `is` or `as`.

Note that while tying the `refine(V, T)` function to the logic of `is` and `as` is necessary, it is not sufficient for at least three reasons:
* There are types that we do not allow to be written as part of type tests (for instance function types) that right now this diff will allow. Since `Typing.refine_and_simplify_intersection` was to written to handled types like this it is likely incorrect. I'll add some additional filtering in a subsequent diff
* While `Typing.refine_and_simplify_intersection` uses existing logic in the type checker, its behavior is not well-defined. Using it as part of the recoverable definition is pragmatic, but for what we release to users I intend to define another relation (based on tags in HHVM and whether they overlap) as a more conservative approximation
* The error messages are not very good. While the above definition of recoverable is useful to detect invalid variants, it does not give any reasons *why* the type cannot be recovered. Defining the relation more precisely will allow us to give better reasons to users.

Reviewed By: andrewjkennedy

Differential Revision: D44945717

fbshipit-source-id: fbdb1c081fae3cf7aebaf5bc107c3b21da9fa2d9