find a better solution for tvars during autocompletion
Summary:
All tvars that are not solved at the end of typing are solved "down" to their lower bounds, even if they appear contravariantly or invariantly. This leads to a lot of tvars solved to `nothing` when we have more information about them, like upper bounds.

This diff tries to be more clever during autocompletion to improve UX

Reviewed By: ljw1004

Differential Revision: D46392004

fbshipit-source-id: cbb501e81c7775459c848ccd5ea0e60e4a959300