Add an index on all_paths(path)
Summary:
I was annoyed waiting for my devserver to warm up, it seemed totally
stuck after a rebase! So I grabbed a profile
and noticed that HHVM was spending a *ton* of time deleting paths.

I don't know anything about SQLite. But some basic googling says that it's
because delete has to rebalance some internal data structure which takes a long
time. And from the look of this query, it might hit an especially-bad case
where it rebalances over and over again.
https://stackoverflow.com/questions/21707190/sqlite-delete-incredibly-slow-how-to-speed-up

Also, if I just skim through some of the queries in sqlite-autoload-db.cpp, it looks like we do a fair amount of matching on all_paths.path, which would lead to full table scans without an index?

We could do the 'pragma' whatever to fit everything into memory, but I think
that might grow the memory use of facts O(repo) which is, in general, not good,
so maybe this is a better approach? Anecdotally my devserver does seem snappier
after I manally add the index to my factdb.

Reviewed By: edwinsmith

Differential Revision: D46382678

fbshipit-source-id: 4e8f35fc0cc50308b73bbedbe65d9d6475dcd920