refactor to extract shape type
Summary:
Refactor to extract shape type

https://pxl.cl/2Kv8N

Theoretically we could do this for *any* type but it's hard to come up with use cases for aliases besides shape aliases.

## Future enhancement opportunities

A potential enhancement is to avoid creating "duplicate" shape types, but given that shapes are structurally typed "duplicate" isn't so well-defined: just because two things have the same shape doesn't mean they are intended to be the same.

As a follow-on (T155067249) we could at least inform the user that the shape is similar to an existing one, but that requires a reliable way to see outside the current file, which is work in progress.

Reviewed By: ljw1004

Differential Revision: D46410691

fbshipit-source-id: d31dcb23d1de4f543a815e8e939fbb83d971b2f0