Treat package change as a major change
Summary:
When a module change changes the package a file belongs to, we should treat the change as a major change and recheck all dependencies of a class.

Note that this change uses the Provider_context to grab the package info object. milliechen is in the process of migrating the package info to use TypecheckerOptions instead of directly inside the context, but since the tcopt is available in the provider context, this change will still be supported in the new API.

Reviewed By: ljw1004

Differential Revision: D46565795

fbshipit-source-id: 375bfd5a2116e11c508a216c38b2f215f9a362de