Handle non-enforceable types when converting from bytecode to IR
Summary:
When the type is not enforceable we get a TypeInfo with an empty type
constraint name in hhbc::TypeInfo. Previously, we'd translate it into a class
constraint with an empty name which is bad. Now we translate such cases as type
info with a base type = None.

Note that we can't simply fold this case under the Mixed base type because it
would break bc -> ir -> bc validation (we need to differentiate between mixed
and empty).

I also removed a TODO in IR -> textual instr lowering around empty class ids.

Reviewed By: aorenste

Differential Revision: D46561378

fbshipit-source-id: 8221d806d9092682a7396ac1951e8db0d98c867f