Fix clap options
Summary:
Trying to convert hh_distc to use cli::main exposed two config problems which are revealed using Clap's debug_assert() in `@//mode/dev` (which is done automatically by cli::main).

First:
clap has some weirdness with a `PathBuf` config value with a default_value of empty string.

When trying to omit the parameter clap will complain:
```
error: a value is required for '--input-dir <INPUT_DIR>' but none was supplied
```
but debug_assert() just outright rejects it.

Removing the bad default_value changes the error to:
```
For more information, try '--help'.
error: the following required arguments were not provided:
  --input-dir <INPUT_DIR>
```
but the behavior is fundamentally the same.
 ---
Second:
The long name "--config" is used in both `HhvmOptions` and `DeclOptions`.  This is fine for most of the commands but for `compile` we flatten both into a single `struct` which causes the long names to collide.

A dbg or dev build will complain if you try to use `compile` but a non-dbg build won't - it just ignores the `DeclOptions` one.  debug_assert() also complains when verifying the options.

As far as I can tell no code actually ever currently reads `HhvmOptions.config_files` so I renamed that option to `--config-file`.  This means we no longer match HHVM - but nothing is currently affected.

Reviewed By: edwinsmith

Differential Revision: D46618397

fbshipit-source-id: 1842b1f7d695c689a963f38a7ec4b56d211cbd8d