type checking for typed local variable initialisation
Summary:
Initial implementation of type checking for typed local variables.

We follow Hack's flow-based typing strategy, rather than implementing a more Java/C-like lexical scoping model. We ensure that within a function, in all control flow subsequent to a typed local declaration, any assignment to the variable is a subtype of the declared type, and furthermore in any re-declaration, the newly declared type must be a subtype of the old one. The actual type of the variable is not affected. For example, the following is accepted.
```
function expect_string(string $x) : void {}

let $x : arraykey = 1;
$x = "";
expect_string($x);
```

To implement this, we add a field to the local variable enviroment tracking what the bound is. When control flow joins, we intersect the bounds to make sure that all of them are followed from that point. We do an additional check that the type still follows the bounds.

Still TODO in subsequent diffs:
  - more extensive testing for refinement, loops, etc.
  - improve error messages, which will require some refactoring of how positions are handled at control flow joins
  - fix the following, which will need us to keep "undefined" locals in the environment
```
  if (...) {
    let $x : int = 1;
  }
  $x = "";
```
This should be rejected since `$x` is being assigned in violation of the bound. Currently, `$x` is not in the local environment, so the assignment is allowed. We need to add a state for locals where the `$x` is undefined so that there is an error on uses, but the bound is still present.

NB, `tast_typecheck.ml` was meant to be deleted in D46446495, but the deletion got moved here in some rebasing.

Reviewed By: vsiles

Differential Revision: D46227210

fbshipit-source-id: ec335c88c103dff3d532f0b0ecc818795edfc021