add heuristic to prevent too many `nothing` during autocompletion
Summary:
Because we don't store the constraints of unsolved tvars for autocompletion, we end up having a lot of `nothing` for invariant/contravariant variable, which disable any efficient filtering of types since we end up trying to check subtype relations where the rhs is `nothing`.

In this diff, I'm testing a heuristic for autocompletion: we no longer "bind to lower" any tvars that would be bound to `nothing`. Doing so, it will fall back to my previous diff, improving the solutions for invariant/contravariant variables. The core idea remains the same "let's not eagerly solve to nothing" as it mostly prevent any further check to succeed.

Differential Revision: D46590416

fbshipit-source-id: b9a6cefd3e711648c7eb69457fa47a0a9a740138