1/6 rewrite ClientIdeInit - simplify delay
Summary:
[This series of 6 diffs rewrites the "load naming table + search index" parts of ClientIdeDaemon. I'm doing this because I want to rollout Tom's "ide_batch_process_changes" feature which changes performance characteristics, but the existing telemetry wasn't adequate for measuring its impact, and the code had grown too complex to tweak.]

This diff is a pretty small change:
```
BEFORE:
   let%lwt () =
      if Float.(naming_table_load_info.test_delay > 0.0) then
         Lwt_unix.sleep naming_table_load_info.test_delay
      else
        Lwt.return_unit

AFTER:
   let%lwt () = Lwt_unix.sleep naming_table_load_info.test_delay in
```
I pulled it out into a separate diff because it's the only behavior-changing alteration in this diffstack. It means that in the case where "test_delay = 0.0" then the load used to happen synchronously, but will now happen on the next Lwt tick (if I understand correctly the implementation of Lwt_unix.sleep).

This feature is only used in tests. If the tests still complete fine with that extra tick of asynchrony, it'll be worth it to save four lines of code!!

Reviewed By: mheiber

Differential Revision: D46616937

fbshipit-source-id: aa3041dd3078715eacedc2afb87954cddcfbb6d1