6/6 rewrite ClientIdeInit - change order of attempts
Summary:
[This series of 6 diffs rewrites the "load naming table + search index" parts of ClientIdeDaemon. I'm doing this because I want to rollout Tom's "ide_batch_process_changes" feature which changes performance characteristics, but the existing telemetry wasn't adequate for measuring its impact, and the code had grown too complex to tweak.]

The existing IDE naming-table-load behavior has an unexpected order of attempts:
1. attempt to find a naming-table on disk
2. otherwise, use the one specified in the LSP initialize method (used primarily for tests)
3. otherwise, attempt to download one
4. otherwise, attempt to build one if allowed by .hhconfig

I think the order of (1) and (2) was the wrong way round. (but prior to the previous diff, the code was convoluted enough that we never even realized!) I think that if a test specifies a naming-table to use in its Initialize method, then by golly we should for sure use it.

This diff changes the order.

Reviewed By: mheiber

Differential Revision: D46624867

fbshipit-source-id: 1e28f7df254c55bba20ff79762b447843b10bd38