4/6 rewrite ClientIdeInit - rationalize errors, stop_reason
Summary:
[This series of 6 diffs rewrites the "load naming table + search index" parts of ClientIdeDaemon. I'm doing this because I want to rollout Tom's "ide_batch_process_changes" feature which changes performance characteristics, but the existing telemetry wasn't adequate for measuring its impact, and the code had grown too complex to tweak.]

The way we did errors out of clientIdeDaemon wasn't serving us well. Look at this code. Why do we have two separate things, `reason` and `e`, that are both kind of the same? This diff removes the redundancy.
```
BEFORE:
    let reason = ClientIdeUtils.make_bug_reason "load_exn" ~exn in
    let e = ClientIdeUtils.make_bug_error "load_exn" ~exn in
    Lwt.return_error (reason, e)
```

BEFORE, `make_bug_error` produced an Lsp.Error.t which we used (1) to return to VSCode in response to its request, (2) stored in clientIdeDaemon Failed_init state, used to indicate that clientIdeDaemon failed to load a naming-table, so it could be returned on all future requests.

`make_bug_reason` produce a ClientIdeMessage.stopped_reason which we used to (1) populate the VSCode status bar and optionally pop up an alert so the user knows they need to take action, (2) store in ClientIdeService.Stopped state -- clientIdeService is the proxy within the clientLsp process that's used to describe its belief about the state of the clientIdeDaemon process, and it needs to retain the stop_reason so it can tell what it thinks should be the contents of the status-bar at any time.

AFTER, I renamed ClientIdeMessage.stopped_reason to `rich_error` and made it a little richer. Previously it was only suitable for rendering to humans. Now I teased apart some of its internal fields so it's also suitable for rendering in Lsp.Error.t messages, and better for scuba telemetry.

Because it's richer, we can now synthesize a Lsp.Error.t out of a `ClientIdeMessage.rich_error` whenever we want. Therefore there's no longer any need for the two of them. And I standardized on storing `rich_error` as part of ClientIdeService state.

I made a few other nearby changes too. I renamed "message" to "category", to be suggestive that this string field will go into scuba where it should have one of a finite number of string values (suitable for GROUP BY), rather than including arbitrary data. I didn't think it was worth the engineering overhead to go full ADT. Also, I renamed some variables from `exn` to `e` following the standard naming elsewhere in our codebase that `exn` is an ocaml exception and `e` is an Exception.t

Reviewed By: mheiber

Differential Revision: D46622773

fbshipit-source-id: fb23257f976c88876f4c8bf0da399d88346c8797