Eval.AutoloadInitEarly - sync autoload db before compiling anything
Summary:
The gold standard for Native autoloading bootstrapping is that the map is initialized and ready before compiling the first line of hack code in any request, including prelude, systemlib, whatever.

We already did this with EnableDecl=true, but we want to test this precondition even without EnableDecl (maybe there are unrelated reasons to not flip EnableDecl yet).

Reviewed By: mdko

Differential Revision: D46546045

fbshipit-source-id: b34b623698b09fdaa2042d1db9fbe0318da63963