Recommend ?? Alongside Shapes::idx()
Summary: Engineers can fix optional shape dict access either by wrapping in Shapes::idx() as is currently recommended or by using ?? to provide a default value for null and undefined. This diff updates the message to indicate that ?? is a valid fix as well (assuming you want the same default to be used for null and undefined).

Differential Revision: D46613761

fbshipit-source-id: deca72c544fd17215b85cc052e471f731aa97dae