Add CLI to Retrieve Snippet for XHP Component
Summary:
We (MDV team) are planning to add Bloks component suggestion through completion service. We're building our own endpoint to retrieve information for all Bloks components in www via Glean. Currently, we're missing the autocomplete snippet, which was added to Hack server earlier this half. Ultimately, we may want to have our own autocomplete snippet baked from the endpoint.

For convenience, we'll be borrowing the autocomplete snippet from Hack server given the name of the component.

Reviewed By: ljw1004

Differential Revision: D46556742

fbshipit-source-id: a14e08d840f644c9450a4a1dd9e1cbcd9aa3a40d