Move Package.Info into its own file
Summary: This is necessary because when we move Package.Info.t into TypecheckerOptions, we will need an oxidized package info, and oxidized_regen.sh doesn't seem to be able to oxidize the module Info within the file package.ml. I believe having a separate file for PackageInfo is also cleaner.

Reviewed By: jamesjwu

Differential Revision: D46539254

fbshipit-source-id: fe80e7d2c21ca94dcd1ada2bea5cde67ebefc5c8