Fix nullptr passed into will_symbol_raise_deployment_boundary_violation
Summary:
If you pass nullptr into
will_symbol_raise_deployment_bound_violation (which is an entirely too
long function name), it will assert. If this function doesn't ever
want to be given a nullptr, then it should take the values by
reference instead. This forces the caller to do the nullptr check and
do the appropriate thing.

Reviewed By: oulgen, jano

Differential Revision: D46693001

fbshipit-source-id: 02687ea300dae9be980112bfc20a47afff4de5c8