improve flow-sensitive checking for typed locals
Summary: Ensure that declared type bounds on locals are checked even after control-flow joins, where a typed local variable is declared in only some of the branches. We do this by keeping such variables in the environment to keep their bounds, but marking them as `undefined` to keep the error if they are accessed.

Reviewed By: vsiles

Differential Revision: D46446190

fbshipit-source-id: e6d371cffbb65aa6f256ee4fe48b7ea9fb0acf32