Fix checking for typed local in loops
Summary:
If a loop introduces a typed local, then we might need to go around again in case it was assigned prior to the definition of the typed local. We do this by making new typed locals not be considered a sub_entry of the previous environment.

When there is an assignment that violates the bound, we set the type of the variable to the bound to check for further errors.

Fix a bug where unbounded assignments (that weren't errors) removed the bound.

Reviewed By: vsiles

Differential Revision: D46720973

fbshipit-source-id: 65d3bf8b83feb167af0042ea372703b273d78296