Auto-rename placeholders in Hack refactors
Summary:
For "extract into variable" and "extract into method" refactors, automatically trigger the "rename" flow for the user.

The previous behavior was to auto-generate a name like `$placeholder1` without triggering the rename flow.

Here's what it looks like:

https://pxl.cl/2HWxs

## Technical choice

The behavior in this diff relies on the client understanding snippet edits.  We *could* check client capabilities and do different behavior depending on client capabilities. Then we could avoid generating weird placeholders for clients that don't understand snippet edits.

I didn't check capabilities so we can have less code to maintain and so the changes are easier to test.

Reviewed By: ljw1004

Differential Revision: D46075794

fbshipit-source-id: eb56082dd859e6d9ae0883beecea5b04448c795f