Encode sequence numbers into request metadata via `pack*`
Summary:
D46044472 summarizes the rationale for sequence numbers. Here, I plumb the relevant `AsyncTransport*` (which is really an `AsyncFdSocket` whenever FDs are present) into Rocket's `pack*` functions.

This lets me encode the socket-generated FD sequence number inside the message's metadata.

A (desirable) side effect of this change is that the send code will now automatically detect if `pack*` happens in a different order than `writeChainWithFds`, and fail that write. In other words, sequence numbers guarantee no message/FD reordering from the `pack*` on the sender all the way through `popReceivedFdsFromSocket` on the receiver. This ensures the right message always has the right FDs.

Reviewed By: AkramaMirza

Differential Revision: D46005254

fbshipit-source-id: 36daf3d39c421360fe39235a62e1ce32e16e831e