StopTLS downgrades Unix sockets to AsyncFdSocket
Summary: Combined with D46364470, this makes it possible to use a StopTLS handshake on Unix connections (which gives you peer certs for authorization) but then send actual requests over an unencrypted `AsyncFdSocket` which avoids the CPU overhead of encryption, and allows FD-passing. I.e. this sets on-host services for the all the functionality they could need -- peer certs, FD-passing, and fast Thrift comms.

Reviewed By: AjanthanAsogamoorthy

Differential Revision: D46360525

fbshipit-source-id: d80cd917c416e7bdfefd0eb836728ab345187199