Search for index and nameIndex in one table lookup
Summary:
The old encode algorithm was:

1. Check to see if this header is indexable in the dynamic table (via strategy)
2. Get the index for the name+value in either the static or dynamic table
3. If found, encode it, done
4. If not, look up the name of the header in the static or dynamic table
5. Encode literal (with optional name reference)

The observation is that the lookup in step 2 necessarily has the answer to the lookup in step 4.  Further, I observed that a powerful HPACK encoding strategy is to choose to index only one entry for a given header name in the dynamic table, and for this, we need to know if the requested header has an index before invoking the indexing strategy.

The new encode algorithm is:

1. Lookup the name and the name/value in the static and dynamic table - getIndex now returns a pair.
2. If there's an exact match, encode it, done.
3. Compute whether the header is indexable
4. Encode the literal with the (optional) name index from 1

Note this allows for referencing a full entry in the static table even if the strategy (erroneously?) would have returned false.

This diff by itself has mixed effects in the encoding performance, but the ability to use the flag in the encoding strategy is very powerful.

Reviewed By: lnicco

Differential Revision: D46669233

fbshipit-source-id: f1badfb1a37f8b0a2bcb668727b7fa68cdbe597f