Also return list of members for a class
Summary:
Return a list of members as well. The list includes name, location, kind and a snippet.

Origin is added to help distinguish inherited vs local members later on in the explorer UI.

Note: In this diff I print `member: type` (e.g. `foo: (function(int):void)`) which is not ideal. I intend to add a pretty printer for declaration in a future diff to print it as `foo(int $param): void`

Reviewed By: mheiber

Differential Revision: D46538133

fbshipit-source-id: eb3a67f872d39b712a28fef94508401a97879c6d